pyproject.toml
[tool.poetry]
name = "initiative-tracker"
version = "0.1.0"
description = ""
authors = ["Dmitry O <feanarokurufinve@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tortoise-orm = "^0.20.0"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
asyncpg = "^0.28.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
